+++ src/main/java/eu/trentorise/smartcampus/mobility/gamificationweb/ReportEmailSender.java
@@ -29,7 +29,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.security.crypto.codec.Base64;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.client.RestTemplate;
 
@@ -292,7 +291,7 @@ public class ReportEmailSender {
 			int point_green_w = 0;
 			if (states != null && states.size() > 0) {
 				if (currentWeekConfData.getWeekStart() != null && currentWeekConfData.getWeekEnd() != null) {
-					point_green = states.get(0).getScore();
+					point_green = (int)states.get(0).getScore();
 					LocalDate cws = LocalDate.parse(currentWeekConfData.getWeekStart());
 					LocalDate cwe = LocalDate.parse(currentWeekConfData.getWeekEnd());
 
@@ -303,7 +302,7 @@ public class ReportEmailSender {
 					}).findFirst().orElse(null);
 
 					if (pcp != null) {
-						point_green_w = pcp.getScore();
+						point_green_w = (int)pcp.getScore();
 					}
 				}
 			}
